backend/requirements.txt
# Core Flask Application
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Werkzeug==2.3.7
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.7

# Data Science & ML Core
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2

# Multi-Spectral Analysis & Visualization
matplotlib==3.7.2
seaborn==0.12.2
scipy==1.11.4

# Satellite & Geospatial Data
requests==2.31.0
earthengine-api==0.1.384
google-auth==2.23.4
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-cloud-storage==2.10.0

# Enhanced ML for Agriculture
lightgbm==4.1.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
python-dotenv==1.0.0
bcrypt==4.0.1

# Image Processing for Visualizations
Pillow==10.0.1

# Optional - Advanced ML libraries (commented out for lighter build)
# torch==2.0.1
# tensorflow==2.13.0
# xgboost==1.7.6
